move command and docs
