+++ site/docs/usage/command-line.md
@@ -35,7 +35,9 @@ The `promptfoo` command line utility supports the following subcommands:
 - `validate` - Validate a promptfoo configuration file.
 - `feedback <message>` - Send feedback to the Promptfoo developers.
 - `import <filepath>` - Import an eval file from JSON format.
-- `export <evalId>` - Export an eval record to JSON format.
+- `export` - Export eval records or logs.
+  - `export eval <evalId>`
+  - `export logs`
 - `redteam` - Red team LLM applications.
   - `redteam init`
   - `redteam setup`
@@ -234,13 +236,28 @@ Deletes a specific resource.
 
 Import an eval file from JSON format.
 
-## `promptfoo export <evalId>`
+## `promptfoo export`
+
+Export eval records or logs.
+
+### `promptfoo export eval <evalId>`
 
 Export an eval record to JSON format. To export the most recent, use evalId `latest`.
 
 | Option                    | Description                                 |
 | ------------------------- | ------------------------------------------- |
 | `-o, --output <filepath>` | File to write. Writes to stdout by default. |
+
+### `promptfoo export logs`
+
+Collect and zip log files for debugging purposes.
+
+| Option                    | Description                                          |
+| ------------------------- | ---------------------------------------------------- |
+| `-n, --count <number>`    | Number of recent log files to include (default: all) |
+| `-o, --output <filepath>` | Output path for the compressed log file              |
+
+This command creates a compressed tar.gz archive containing your promptfoo log files, making it easy to share them for debugging purposes. If no output path is specified, it will generate a timestamped filename automatically.
 
 ## `promptfoo validate`
